partially extract the selection tool
this partial extraction makes it possible to put the selection
tool in the drawing tile's tool registry

TODO: move more of the selection specific methods out
of the drawing-layer like the selection box.
