+++ src/plugins/drawing-tool/components/drawing-layer.tsx
@@ -12,7 +12,7 @@ import { ImageContentSnapshotOutType } from "../../../models/tools/image/image-c
 import { gImageMap } from "../../../models/image-map";
 import { SelectionBox } from "./selection-box";
 import { DrawingObjectSnapshot, DrawingObjectType, DrawingTool, 
-  HandleObjectHover, IDrawingLayer } from "../objects/drawing-object";
+  HandleObjectHover } from "../objects/drawing-object";
 import { Point, ToolbarSettings } from "../model/drawing-basic-types";
 import { applyAction, getMembers, getSnapshot, SnapshotOut } from "mobx-state-tree";
 import { DrawingObjectMSTUnion,
@@ -25,54 +25,6 @@ const SELECTION_BOX_PADDING = 10;
 
 function makeSetter(prop: string) {
   return "set" + prop.charAt(0).toUpperCase() + prop.slice(1);
-}
-
-/**  ======= Drawing Tools ======= */
-class SelectionDrawingTool extends DrawingTool {
-  constructor(drawingLayer: IDrawingLayer) {
-    super(drawingLayer);
-  }
-
-  public handleMouseDown(e: React.MouseEvent<HTMLDivElement>) {
-    // We are internal so we can use some private stuff not exposed by 
-    // IDrawingLayer
-    const drawingLayerView = this.drawingLayer as DrawingLayerView;
-    const addToSelectedObjects = e.ctrlKey || e.metaKey;
-    const start = this.drawingLayer.getWorkspacePoint(e);
-    if (!start) return;
-    drawingLayerView.startSelectionBox(start);
-
-    const handleMouseMove = (e2: MouseEvent) => {
-      e2.preventDefault();
-      const p = this.drawingLayer.getWorkspacePoint(e2);
-      if (!p) return;
-      drawingLayerView.updateSelectionBox(p);
-    };
-    const handleMouseUp = (e2: MouseEvent) => {
-      e2.preventDefault();
-      drawingLayerView.endSelectionBox(addToSelectedObjects);
-      window.removeEventListener("mousemove", handleMouseMove);
-      window.removeEventListener("mouseup", handleMouseUp);
-    };
-
-    window.addEventListener("mousemove", handleMouseMove);
-    window.addEventListener("mouseup", handleMouseUp);
-  }
-
-  public handleObjectClick(e: React.MouseEvent<HTMLDivElement>, obj: DrawingObjectType) {
-    // We are internal so we can use some private stuff not exposed by 
-    // IDrawingLayer
-    const drawingLayerView = this.drawingLayer as DrawingLayerView;
-    const {selectedObjects} = drawingLayerView.state;
-    const index = selectedObjects.indexOf(obj);
-    if (index === -1) {
-      selectedObjects.push(obj);
-    }
-    else {
-      selectedObjects.splice(index, 1);
-    }
-    drawingLayerView.setSelectedObjects(selectedObjects);
-  }
 }
 
 /**  ======= Drawing Layer ======= */
@@ -122,15 +74,13 @@ export class DrawingLayerView extends React.Component<DrawingLayerViewProps, Dra
       hoverObject: null,
     };
 
-    this.tools = {
-      selection: new SelectionDrawingTool(this)
-    };
+    this.tools = {};
     const drawingToolInfos = getDrawingToolInfos();
     drawingToolInfos.forEach(info => {
       this.tools[info.name] = new info.toolClass(this);
     });
 
-    this.currentTool = this.tools.selection!;
+    this.currentTool = this.tools.select!;
 
     this.objects = {};
 
@@ -178,9 +128,7 @@ export class DrawingLayerView extends React.Component<DrawingLayerViewProps, Dra
 
   public syncCurrentTool(selectedButton: string) {
     const settings = this.getContent().toolbarSettings;
-
-    const toolName = selectedButton === "select" ? "selection" : selectedButton;
-    const tool = this.tools[toolName];
+    const tool = this.tools[selectedButton];
     
     if (!tool) {
       console.warn("Unknown tool selected", selectedButton);
@@ -254,7 +202,7 @@ export class DrawingLayerView extends React.Component<DrawingLayerViewProps, Dra
   };
 
   public handleObjectHover: HandleObjectHover = (e, obj, hovering) => {
-    if (!this.props.readOnly && this.currentTool === this.tools.selection) {
+    if (!this.props.readOnly && this.currentTool === this.tools.select) {
       this.setState({hoverObject: hovering ? obj : null});
     }
   };
